requirements.txt
fastapi
gunicorn
pytest
python-dotenv
uvicorn
coverage
pydantic-settings